relocating files, fixed signup routing
