+++ exam-machine/src/app/dashboard/mentor/students/students.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {DataService } from '../../../data.service';
-import { AlertService } from '../../../alert.service';
+import {DataService } from '../../../services/data.service';
+import { AlertService } from '../../../services/alert.service';
 import 'rxjs/add/operator/toPromise';
 
 @Component({
