--- exam-machine/src/app/loading/loading.component.css
@@ -1,8 +1,0 @@
-section {
-  width: auto;
-  height: calc(100vh - 50px);
-  display: flex;
-  align-items: center;
-  align-content: center;
-  justify-content: center;
-}
