--- exam-machine/src/app/loading/loading.component.html
@@ -1,5 +1,0 @@
-<section>
-  <div class="spinner spinner-inverse">
-    Loading...
-  </div>
-</section>
