+++ exam-machine/src/app/screens/login/login.component.html
@@ -13,7 +13,7 @@
 </section>
 
 <section class="alert-container" [ngStyle]="{'height': height}">
-  <div [ngClass]="{'alert-danger': dangerAlert, 'alert-success': successAlert}" class="alert">
+  <div [ngClass]="{'alert-danger': errorAlert, 'alert-success': successAlert}" class="alert">
     <div class="alert-item">
       <span class="alert-text">
         {{message}}
