+++ exam-machine/src/app/screens/signup/signup.component.css
