+++ exam-machine/src/app/screens/signup/signup.component.html
@@ -15,8 +15,8 @@
   <!-- <button type="submit" class="btn btn-primary" (click)="hideAlert = !hideAlert;">Sign up</button> -->
 </section>
 
-<section #signupAlert [ngClass]="{'hide': hideAlert}">
-  <div [ngClass]="{'alert-danger': dangerAlert, 'alert-success': successAlert}" class="alert">
+<section class="alert-container" [ngStyle]="{'height': height}">
+  <div [ngClass]="{'alert-danger': errorAlert, 'alert-success': successAlert}" class="alert">
     <div class="alert-item">
       <span class="alert-text">
         {{message}}
