+++ exam-machine/src/app/screens/unauthorized/unauthorized.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { LoginService } from './../login.service';
+import { LoginService } from '../../services/login.service';
 
 @Component({
   selector: 'app-unauthorized',
