+++ exam-machine/src/app/services/activate.service.spec.ts
