+++ exam-machine/src/app/services/activate.service.ts
