+++ exam-machine/src/app/services/calculate.service.spec.ts
