+++ exam-machine/src/app/services/calculate.service.ts
